+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Don't require token for `--dry-run` of `compote component upload` command
 - Fix incorrect message suggestion to check upload status if non-default profile is used
 - Add support for non-default registries to `compote project create-from-example` command through the `--service-profile` option